+++ tests/spec/parser/mod.rs
@@ -4,3 +4,5 @@
 use super::runner;
 
 mod indentation;
+
+mod selector;